Created Schedule List Component
